refactor: split store.go (1625 lines) into domain-focused files (#565)
Pure file reorganization — no logic, signature, or behavior changes.
store.go retains core types, constructor, migrations, transaction helpers,
and type converters (~340 lines). Domain methods moved to:

- quota.go, objects.go, replication.go, multipart.go, usage.go,
  integrity.go, advisory_lock.go, admin.go, notifications.go

Also renames free-tier-5-cloud-setup.png to multi-cloud-setup.png and
updates the new dashboard screenshot.
