+++ .version
@@ -1,1 +1,1 @@
-v0.40.1
+v0.40.2
